Document the router.
The AI did not test the integration hub. I will try for it to do
it mocking the dependencies. I need some tests to check the routes
are correct and the proper UIs are being invoked. It is like testing
protocols, we need to make sure the decision tree and action order
is correct.