Refactor config.rs to accept constructor based injection.
I think I made a mistake here. it seems this struct was all static
methods and I did not realize. So I wanted constructor-based
dependency injeciton and I ended up with more testable code and
ergonomics for the production code at the expense of a much more
complicated code base.

I wonder if there is other opetions to accept dependencies but
still rely on a simplified code base.